Refactor elytra -> fall flying
