+++ fabric-item-api-v1/src/main/resources/fabric.mod.json
@@ -17,7 +17,7 @@
   ],
   "mixins": [
     "fabric-item-api-v1.mixins.json",
-    "fabric-item-api-v1-elytra.mixins.json"
+    "fabric-item-api-v1-fall-flying.mixins.json"
   ],
   "depends": {
     "fabricloader": ">=0.4.0",
