+++ fabric-item-api-v1/src/testmod/resources/fabric.mod.json
@@ -12,7 +12,7 @@
     "main": [
       "net.fabricmc.fabric.test.item.FabricItemSettingsTests",
       "net.fabricmc.fabric.test.item.CustomDamageTest",
-      "net.fabricmc.fabric.test.item.FabricElytraTests"
+      "net.fabricmc.fabric.test.item.FabricFallFlyingTests"
     ],
     "client": [
       "net.fabricmc.fabric.test.item.client.TooltipTests"
